# api/openapi.yaml
openapi: 3.0.3
info:
  title: GitStack API
  version: "0.1.0"
  description: >
    API pubblica di GitStack (fonte unica del contratto, decisione D8
    [c_f68d7615da93197a]). Unico punto d'ingresso: il servizio `gateway`.
    UI web, CLI `gs`, skills per agenti e in futuro server MCP usano tutti
    questa stessa API (principio API-first, decisione D7 [c_3780bc749c6ed28d]).

    Il modello dati usa risorse generiche (decisione D15
    [c_4ef209e79c5b2ddb]): oggi la sola risorsa di prova dello scheletro
    end-to-end, domani repository, applicazioni, database e altro senza
    cambiare forma dell'API.

    Ogni versione dell'API vive sotto un prefisso di percorso esplicito
    (`/v1`); `info.version` segue in aggiunta il contratto stesso (SemVer).
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
...
servers:
  - url: /v1
    description: Gateway pubblico di GitStack (dietro Traefik)
...
paths:
...
  /user/tokens:
    get:
      operationId: listTokens
      tags: [tokens]
      summary: Elenca i token personali dell'utente corrente
      description: Scope `read:user`. Mai il valore del token, solo `hint`.
      x-required-scopes: ["read:user"]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: Pagina di token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/UnexpectedError"
    post:
      operationId: createToken
      tags: [tokens]
      summary: Crea un token personale
      description: >
        Scope `write:user`. Il valore (`gst_...`) e' nella risposta una volta
        sola: il servizio conserva solo l'hash. Un token non puo' avere scope
        che il suo creatore non ha.
      x-required-scopes: ["write:user"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTokenInput"
      responses:
        "201":
          description: Token creato.
          headers:
            Location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedToken"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /user/tokens/{tokenId}:
    parameters:
      - $ref: "#/components/parameters/TokenIdParam"
    delete:
      operationId: revokeToken
      tags: [tokens]
      summary: Revoca un token personale
      description: Scope `write:user`. La revoca e' immediata per identity; il gateway la vede entro il TTL della cache di verifica.
      x-required-scopes: ["write:user"]
      responses:
        "204":
          description: Token revocato.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /user/ssh-keys:
    get:
      operationId: listSshKeys
      tags: [ssh-keys]
      summary: Elenca le chiavi SSH dell'utente corrente
      x-required-scopes: ["read:user"]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: Pagina di chiavi.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SshKeyList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/UnexpectedError"
    post:
      operationId: addSshKey
      tags: [ssh-keys]
      summary: Aggiunge una chiave SSH
      description: >
        Scope `write:user`. Il servizio valida la chiave, calcola il
        fingerprint SHA256 e lo tiene unico: la stessa chiave non puo'
        appartenere a due utenti (409 `ssh_key_in_use`).
      x-required-scopes: ["write:user"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddSshKeyInput"
      responses:
        "201":
          description: Chiave aggiunta.
          headers:
            Location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SshKey"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /user/ssh-keys/{keyId}:
    parameters:
      - $ref: "#/components/parameters/SshKeyIdParam"
    get:
      operationId: getSshKey
      tags: [ssh-keys]
      summary: Legge una chiave SSH
      x-required-scopes: ["read:user"]
      responses:
        "200":
          description: Chiave trovata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SshKey"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/UnexpectedError"
    delete:
      operationId: deleteSshKey
      tags: [ssh-keys]
      summary: Elimina una chiave SSH
      x-required-scopes: ["write:user"]
      responses:
        "204":
          description: Chiave eliminata.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/UnexpectedError"
...
  /internal/verify:
    post:
      operationId: verifyCredential
      tags: [internal]
      summary: Verifica una credenziale (per il gateway)
      description: >
        Il gateway invia il valore grezzo della credenziale (cookie di
        sessione oppure token `gst_...`) e riceve il principal. Non c'e'
        distinzione fra "sconosciuta", "scaduta" e "revocata": `active` false.
        Aggiorna `last_used_at` (al piu' una volta al minuto per credenziale).
      security:
        - serviceAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCredentialInput"
      responses:
        "200":
          description: Esito della verifica.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyCredentialResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /internal/permissions/check:
    post:
      operationId: checkPermission
      tags: [internal]
      summary: Verifica un permesso su una risorsa
      description: >
        Usata da core e git per applicare i permessi: `allowed` e' vero se
        l'utente ha almeno il ruolo richiesto sulla risorsa (grant diretto,
        via team o amministratore).
      security:
        - serviceAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckPermissionInput"
      responses:
        "200":
          description: Esito.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckPermissionResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /internal/ssh-keys/{fingerprint}:
    parameters:
      - name: fingerprint
        in: path
        required: true
        description: Fingerprint `SHA256:...` (URL-encoded).
        schema:
          type: string
          pattern: "^SHA256:[A-Za-z0-9+/]{43}$"
    get:
      operationId: lookupSshKey
      tags: [internal]
      summary: Risolve un utente dal fingerprint di una chiave SSH
      description: Usata dal servizio git all'accesso SSH. Aggiorna `lastUsedAt`.
      security:
        - serviceAuth: []
      responses:
        "200":
          description: Chiave e utente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SshKeyLookup"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/UnexpectedError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: >
        Token personale `gst_...` verificato dal `gateway` tramite identity.
        Stesso schema per CLI `gs`, skills e agenti. Ogni operazione di
        identity indica gli scope richiesti in `x-required-scopes`; le
        sessioni web (cookie) non hanno scope.
    sessionCookie:
      type: apiKey
      in: cookie
      name: gst_session
      description: >
        Sessione web creata da `POST /auth/login` o dal login OIDC. Valore
        opaco; identity ne conserva solo l'hash.
    serviceAuth:
      type: http
      scheme: bearer
      description: >
        Solo interfaccia interna: segreto condiviso di servizio fra gateway,
        core, git e identity. Non e' un token utente.
  parameters:
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: perPage
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    TokenIdParam:
      name: tokenId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SshKeyIdParam:
      name: keyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Error:
      type: object
      description: >
        Formato unico degli errori per tutta l'API pubblica. Codici comuni per
        stato: 400 `bad_request`; 401 `unauthenticated`, `invalid_credentials`;
        403 `forbidden`, `insufficient_scope`, `password_change_required`; 429 `too_many_attempts`; 404 `not_found`; 409 `conflict`
        e varianti specifiche (`already_exists`, `last_admin`, `last_owner`,
        `ssh_key_in_use`, `oidc_identity_unlinked`); 422 `validation_failed`
        (`details.fields`). Le risposte 401 non distinguono mai "utente
        inesistente" da "password errata".
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              description: Codice stabile leggibile da macchina (es. `not_found`).
              example: not_found
            message:
              type: string
              description: Messaggio leggibile da persone, non contrattuale.
            details:
              type: object
              additionalProperties: true
              description: Dettagli opzionali specifici dell'errore.
    Name:
      type: string
      description: >
        Nome breve in minuscolo (username, organizzazione, team, provider):
        lettere minuscole, cifre e trattini, 1-39 caratteri, inizia e finisce
        con un carattere alfanumerico.
      pattern: "^[a-z0-9]([a-z0-9-]{0,37}[a-z0-9])?$"
      minLength: 1
      maxLength: 39
      example: alice
    TokenScope:
      type: string
      description: >
        Catalogo degli scope dei token personali (unico, condiviso da tutti i
        servizi). Le sessioni web non hanno scope: valgono i permessi
        dell'utente. `write:*` include `read:*` dello stesso ambito;
        `admin:org` include `write:org`.
      enum:
        - read:user
        - write:user
        - read:org
        - write:org
        - admin:org
        - read:resource
        - write:resource
    ResourceRole:
      type: string
      description: Ruolo su una risorsa, in ordine crescente di potere.
      enum: [read, write, admin]
...
    User:
      type: object
      description: >
        Profilo utente. `email`, `isAdmin`, `isActive`, `createdAt` sono
        presenti solo per l'utente stesso e per gli amministratori.
      additionalProperties: false
      required: [id, username, kind, displayName]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        username:
          $ref: "#/components/schemas/Name"
        kind:
          type: string
          enum: [human, agent]
        displayName:
          type: string
          maxLength: 128
        bio:
          type: string
          maxLength: 1024
        avatarUrl:
          type: string
          format: uri
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        isAdmin:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
          readOnly: true
...
    Token:
      type: object
      description: Token personale, senza il valore (che non e' recuperabile).
      additionalProperties: false
      required: [id, name, scopes, hint, createdAt]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          minLength: 1
          maxLength: 64
        scopes:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/TokenScope"
        hint:
          type: string
          minLength: 4
          maxLength: 4
          description: Ultimi 4 caratteri del token, per riconoscerlo.
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
    CreateTokenInput:
      type: object
      additionalProperties: false
      required: [name, scopes]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        scopes:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/TokenScope"
        expiresAt:
          type: string
          format: date-time
          description: >
            Obbligatoria: la scadenza deve essere nel futuro e non oltre il
            massimo configurato dall'installazione (default 365 giorni), altrimenti 422.
    CreatedToken:
      description: Il token appena creato; `token` compare solo in questa risposta.
      allOf:
        - $ref: "#/components/schemas/Token"
        - type: object
          required: [token]
          properties:
            token:
              type: string
              pattern: "^gst_[A-Za-z0-9_-]{32,}$"
              description: Valore in chiaro, formato `gst_<casuale>`. Non viene salvato.
    TokenList:
      type: object
      additionalProperties: false
      required: [items, page, perPage, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Token"
        page:
          type: integer
          minimum: 1
        perPage:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
    SshKey:
      type: object
      additionalProperties: false
      required: [id, title, keyType, fingerprint, createdAt]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
          minLength: 1
          maxLength: 64
        keyType:
          type: string
          example: ssh-ed25519
        fingerprint:
          type: string
          pattern: "^SHA256:[A-Za-z0-9+/]{43}$"
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
    AddSshKeyInput:
      type: object
      additionalProperties: false
      required: [title, publicKey]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 64
        publicKey:
          type: string
          maxLength: 8192
          example: "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAA... alice@laptop"
          description: Riga in formato `authorized_keys` (tipo, base64, commento facoltativo).
    SshKeyList:
      type: object
      additionalProperties: false
      required: [items, page, perPage, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SshKey"
        page:
          type: integer
          minimum: 1
        perPage:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
    SshKeyLookup:
      type: object
      additionalProperties: false
      required: [key, user]
      properties:
        key:
          $ref: "#/components/schemas/SshKey"
        user:
          $ref: "#/components/schemas/User"
...
    VerifyCredentialInput:
      type: object
      additionalProperties: false
      required: [credential]
      properties:
        credential:
          type: string
          minLength: 1
          maxLength: 512
          writeOnly: true
          description: Valore grezzo della credenziale (cookie `gst_session` o token `gst_...`).
        kind:
          type: string
          enum: [session, token]
          description: Facoltativo, evita l'ambiguita' se il gateway la conosce.
    Principal:
      type: object
      description: Chi ha presentato la credenziale.
      additionalProperties: false
      required: [userId, username, kind, isAdmin, authMethod]
      properties:
        userId:
          type: string
          format: uuid
        username:
          $ref: "#/components/schemas/Name"
        kind:
          type: string
          enum: [human, agent]
        isAdmin:
          type: boolean
        authMethod:
          type: string
          enum: [password, oidc, token]
        scopes:
          type: array
          description: Solo per i token; assenti per le sessioni.
          items:
            $ref: "#/components/schemas/TokenScope"
        credentialId:
          type: string
          format: uuid
          description: Id della sessione o del token (per audit).
        expiresAt:
          type: string
          format: date-time
          nullable: true
    VerifyCredentialResult:
      type: object
      additionalProperties: false
      required: [active]
      properties:
        active:
          type: boolean
        principal:
          $ref: "#/components/schemas/Principal"
        cacheTtlSeconds:
          type: integer
          minimum: 0
          description: Per quanto tempo il gateway puo' tenere in cache l'esito.
    CheckPermissionInput:
      type: object
      additionalProperties: false
      required: [userId, resourceId, role]
      properties:
        userId:
          type: string
          format: uuid
        resourceId:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/ResourceRole"
    CheckPermissionResult:
      type: object
      additionalProperties: false
      required: [allowed]
      properties:
        allowed:
          type: boolean
        effectiveRole:
          type: string
          allOf:
            - $ref: "#/components/schemas/ResourceRole"
          nullable: true
  responses:
...
    BadRequest:
      description: Richiesta non valida.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Token mancante o non valido.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Risorsa non trovata.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflitto con lo stato attuale della risorsa.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnexpectedError:
      description: Errore imprevisto.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: >
        Autenticato ma non autorizzato: scope del token insufficiente
        (`insufficient_scope`, con `details.required`) o permesso mancante
        (`forbidden`), oppure password iniziale ancora da cambiare
        (`password_change_required`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnprocessableEntity:
      description: >
        Richiesta ben formata ma non valida semanticamente (`validation_failed`,
        con `details.fields` = mappa campo -> motivo).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
security:
  - bearerAuth: []
  - sessionCookie: []